refactor: combine net.go with rollback.go
